Rollback movement of monolithic class
