+++ src/buildercore/cfngen.py
@@ -16,7 +16,7 @@ We want to add an external volume to an EC2 instance to increase available space
 """
 import os, json, copy
 import re
-from collections import OrderedDict
+from collections import OrderedDict, namedtuple
 import netaddr
 from . import utils, cloudformation, terraform, core, project, context_handler
 from .utils import ensure, lmap
@@ -399,6 +399,31 @@ UPDATABLE_TITLE_PATTERNS = ['^CloudFront.*', '^ElasticLoadBalancer.*', '^EC2Inst
 REMOVABLE_TITLE_PATTERNS = ['^CloudFront.*', '^CnameDNS\\d+$', 'FastlyDNS\\d+$', '^ExtDNS$', '^ExtraStorage.+$', '^MountPoint.+$', '^.+Queue$', '^EC2Instance.+$', '^IntDNS.*$', '^ElastiCache.*$', '^.+Topic$']
 EC2_NOT_UPDATABLE_PROPERTIES = ['ImageId', 'Tags', 'UserData']
 
+# CloudFormation is nicely chopped up into:
+# * what to add
+# * what to modify
+# * what to remove
+# What we see here is the new Terraform generated.tf file, containing all resources (just Fastly so far).
+# We can do a diff with the current one which would already be an improvement, but ultimately the source of truth
+# is changing it and running a terraform plan to see proposed changes. We should however roll it back if the user
+# doesn't confirm.
+class Delta(namedtuple('Delta', ['plus', 'edit', 'minus', 'terraform'])):
+    """represents a delta between and old and new CloudFormation generated template, showing which resources are being added, updated, or removed
+
+    Extends the namedtuple-generated class to add custom methods."""
+    @property
+    def non_empty(self):
+        return any([
+            self.plus['Resources'],
+            self.plus['Outputs'],
+            self.edit['Resources'],
+            self.edit['Outputs'],
+            self.minus['Resources'],
+            self.minus['Outputs'],
+            self.terraform
+        ])
+_empty_cloudformation_dictionary = {'Resources': {}, 'Outputs': {}}
+Delta.__new__.__defaults__ = (_empty_cloudformation_dictionary, _empty_cloudformation_dictionary, _empty_cloudformation_dictionary, None)
 
 def template_delta(context):
     """given an already existing template, regenerates it and produces a delta containing only the new resources.
@@ -484,7 +509,7 @@ def template_delta(context):
     delta_minus_resources = {r: v for r, v in old_template['Resources'].items() if r not in template['Resources'] and _title_is_removable(r)}
     delta_minus_outputs = {o: v for o, v in old_template.get('Outputs', {}).items() if o not in template.get('Outputs', {})}
 
-    return cloudformation.CloudFormationDelta(
+    return Delta(
         {
             'Resources': delta_plus_resources,
             'Outputs': delta_plus_outputs,
